Move Isolate C++ definitions to isolate.cc
