+++ isolate.cc
@@ -1,0 +1,98 @@
+#include "isolate.h"
+#include "context.h"
+#include "deps/include/v8-context.h"
+#include "deps/include/v8-initialization.h"
+#include "deps/include/v8-locker.h"
+#include "deps/include/v8-platform.h"
+#include "libplatform/libplatform.h"
+
+using namespace v8;
+
+auto default_platform = platform::NewDefaultPlatform();
+ArrayBuffer::Allocator* default_allocator;
+
+extern "C" {
+
+/********** Isolate **********/
+
+#define ISOLATE_SCOPE(iso)           \
+  Locker locker(iso);                \
+  Isolate::Scope isolate_scope(iso); \
+  HandleScope handle_scope(iso);
+
+#define ISOLATE_SCOPE_INTERNAL_CONTEXT(iso) \
+  ISOLATE_SCOPE(iso);                       \
+  m_ctx* ctx = isolateInternalContext(iso);
+
+void Init() {
+#ifdef _WIN32
+  V8::InitializeExternalStartupData(".");
+#endif
+  V8::InitializePlatform(default_platform.get());
+  V8::Initialize();
+
+  default_allocator = ArrayBuffer::Allocator::NewDefaultAllocator();
+  return;
+}
+
+IsolatePtr NewIsolate() {
+  Isolate::CreateParams params;
+  params.array_buffer_allocator = default_allocator;
+  Isolate* iso = Isolate::New(params);
+  Locker locker(iso);
+  Isolate::Scope isolate_scope(iso);
+  HandleScope handle_scope(iso);
+
+  iso->SetCaptureStackTraceForUncaughtExceptions(true);
+
+  // Create a Context for internal use
+  m_ctx* ctx = new m_ctx;
+  ctx->ptr.Reset(iso, Context::New(iso));
+  ctx->iso = iso;
+  iso->SetData(0, ctx);
+
+  return iso;
+}
+
+void IsolatePerformMicrotaskCheckpoint(IsolatePtr iso) {
+  ISOLATE_SCOPE(iso)
+  iso->PerformMicrotaskCheckpoint();
+}
+
+void IsolateDispose(IsolatePtr iso) {
+  if (iso == nullptr) {
+    return;
+  }
+  ContextFree(isolateInternalContext(iso));
+
+  iso->Dispose();
+}
+
+void IsolateTerminateExecution(IsolatePtr iso) {
+  iso->TerminateExecution();
+}
+
+int IsolateIsExecutionTerminating(IsolatePtr iso) {
+  return iso->IsExecutionTerminating();
+}
+
+IsolateHStatistics IsolationGetHeapStatistics(IsolatePtr iso) {
+  if (iso == nullptr) {
+    return IsolateHStatistics{0};
+  }
+  v8::HeapStatistics hs;
+  iso->GetHeapStatistics(&hs);
+
+  return IsolateHStatistics{hs.total_heap_size(),
+                            hs.total_heap_size_executable(),
+                            hs.total_physical_size(),
+                            hs.total_available_size(),
+                            hs.used_heap_size(),
+                            hs.heap_size_limit(),
+                            hs.malloced_memory(),
+                            hs.external_memory(),
+                            hs.peak_malloced_memory(),
+                            hs.number_of_native_contexts(),
+                            hs.number_of_detached_contexts()};
+}
+}
