+++ v8go.cc
@@ -20,9 +20,6 @@
 
 using namespace v8;
 
-auto default_platform = platform::NewDefaultPlatform();
-ArrayBuffer::Allocator* default_allocator;
-
 const int ScriptCompilerNoCompileOptions = ScriptCompiler::kNoCompileOptions;
 const int ScriptCompilerConsumeCodeCache = ScriptCompiler::kConsumeCodeCache;
 const int ScriptCompilerEagerCompile = ScriptCompiler::kEagerCompile;
@@ -87,82 +84,6 @@ extern "C" {
 #define ISOLATE_SCOPE_INTERNAL_CONTEXT(iso) \
   ISOLATE_SCOPE(iso);                       \
   m_ctx* ctx = isolateInternalContext(iso);
-
-void Init() {
-#ifdef _WIN32
-  V8::InitializeExternalStartupData(".");
-#endif
-  V8::InitializePlatform(default_platform.get());
-  V8::Initialize();
-
-  default_allocator = ArrayBuffer::Allocator::NewDefaultAllocator();
-  return;
-}
-
-IsolatePtr NewIsolate() {
-  Isolate::CreateParams params;
-  params.array_buffer_allocator = default_allocator;
-  Isolate* iso = Isolate::New(params);
-  Locker locker(iso);
-  Isolate::Scope isolate_scope(iso);
-  HandleScope handle_scope(iso);
-
-  iso->SetCaptureStackTraceForUncaughtExceptions(true);
-
-  // Create a Context for internal use
-  m_ctx* ctx = new m_ctx;
-  ctx->ptr.Reset(iso, Context::New(iso));
-  ctx->iso = iso;
-  iso->SetData(0, ctx);
-
-  return iso;
-}
-
-static inline m_ctx* isolateInternalContext(Isolate* iso) {
-  return static_cast<m_ctx*>(iso->GetData(0));
-}
-
-void IsolatePerformMicrotaskCheckpoint(IsolatePtr iso) {
-  ISOLATE_SCOPE(iso)
-  iso->PerformMicrotaskCheckpoint();
-}
-
-void IsolateDispose(IsolatePtr iso) {
-  if (iso == nullptr) {
-    return;
-  }
-  ContextFree(isolateInternalContext(iso));
-
-  iso->Dispose();
-}
-
-void IsolateTerminateExecution(IsolatePtr iso) {
-  iso->TerminateExecution();
-}
-
-int IsolateIsExecutionTerminating(IsolatePtr iso) {
-  return iso->IsExecutionTerminating();
-}
-
-IsolateHStatistics IsolationGetHeapStatistics(IsolatePtr iso) {
-  if (iso == nullptr) {
-    return IsolateHStatistics{0};
-  }
-  v8::HeapStatistics hs;
-  iso->GetHeapStatistics(&hs);
-
-  return IsolateHStatistics{hs.total_heap_size(),
-                            hs.total_heap_size_executable(),
-                            hs.total_physical_size(),
-                            hs.total_available_size(),
-                            hs.used_heap_size(),
-                            hs.heap_size_limit(),
-                            hs.malloced_memory(),
-                            hs.external_memory(),
-                            hs.peak_malloced_memory(),
-                            hs.number_of_native_contexts(),
-                            hs.number_of_detached_contexts()};
-}
 
 RtnUnboundScript IsolateCompileUnboundScript(IsolatePtr iso,
                                              const char* s,
